requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
greenlet==1.1.2
idna==3.3
lxml==4.9.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
requests==2.28.1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
urllib3==1.26.10
vk-api==11.9.9
